+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
-## [0.2.0] - 2022-2-25
+## [0.3.0] - 2022-03-01
+### Fixes
+- Fixed deserialization bug where chars that shouldn't be accepted were sometimes accepted when the
+  radix was above 10
+
+### Changes
+- Deserialization uses `from_dec_or_hex_str_restricted`
+
+### Additions
+- added `from_dec_or_hex_str_restricted`
+
+## [0.2.1] - 2022-02-28
+### Fixes
+- Fixed formatting impls to use `pad_integral`. The non-alternate form of `LowerHex` (the one
+  involving the "{:x}" format string) now does not include the leading "0x". Only the alternate
+  "{:#x}" format should include the leading "0x".
+
+## [0.2.0] - 2022-02-25
 ### Fixes
 - Casting is now correct on big endian systems